01_Material_Implementaion/Script/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
langfuse>=3.0.0,<5.0.0
python-dotenv>=1.0.0
ulid-py>=1.1.0